requirements.txt
googleapis-common-protos==1.5.3
grpcio==1.12.1
grpcio-tools==1.12.1